requirements.txt
argon2-cffi==20.1.0
asgiref==3.2.10
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
dj-database-url==0.5.0
Django==3.1.4
django-ckeditor==6.0.0
django-cleanup==2.1.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-js-asset==1.2.2
grep==0.3.2
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2
